Hack a bug that is seen without printf.
Without using printf we get caught in the loop that looks for the
logical controller configuration packet for some reason.  This tells me
that either the read call from the RF24 library isn't synchronous, or
there is some sort of race condition between controllers, interrupts,
etc.

This should be investigated further because this isn't really a "fix",
instead a "hack" that seems to work around whatever issue is there.

The delay amount was determined by trial and error using 3 RF1s and a
string of GECEs on the receiver node.