Fix #22 and $32 Hopefully
Modified the serial pritnting to use progmem,  tested with debug on and
off and verified working on a miniren with kombee controller.   also
removed the renard controller completely and removed the printf library
to rely on a dependency of it existing in the RF24 code which the
RFPixelControl code is kinda dependent on already.