Implement a now-tested "DUMB_BLINKER" feature.
This is a feature for the kömblinkin's "RF_In_Strobe_Out" sketch that
will not use any RF data.  It will simply blink when the sketch is
active.  The idea being that instead of controlling the strobes by
setting an RF channel high, you would plug the Kömblinkin's power supply
into a controlled channel (either a 12V DC controlled channel, or a 120V
channel using a smaller 12V power supply) and enable/disable the strobes
by toggling that channel.

This is useful because the Kömblinkin hardware is a relatively cheap
dedicated strobe controller making it useful for people not actually
using other Komby RF gear.